Support both sync and async callback in `AstPath#{call,callParent}()`
